update related memory instructions

+++ include/common/errinfo.h
@@ -241,12 +241,12 @@ struct InfoInstruction {
 struct InfoBoundary {
   InfoBoundary() = delete;
   InfoBoundary(
-      const uint64_t Off, const uint32_t Len = 0,
+      const uint64_t Off, const uint64_t Len = 0,
       const uint64_t Lim = std::numeric_limits<uint64_t>::max()) noexcept
       : Offset(Off), Size(Len), Limit(Lim) {}
 
   uint64_t Offset;
-  uint32_t Size;
+  uint64_t Size;
   uint64_t Limit;
 };
